Extract RootItem base class
Create new RootItem class containing root item functionality from QueryTree,
making QueryTree extend RootItem.
Also extract MockItem and FakeWSItem from test cases to correct package.
